+++ libs/UpdatePackage.php
@@ -2,6 +2,7 @@
 
 namespace humhub\modules\updater\libs;
 
+use humhub\widgets\Link;
 use Yii;
 use yii\base\Exception;
 use yii\helpers\Json;
@@ -48,6 +49,30 @@ class UpdatePackage
         }
 
         $this->fileName = $fileName;
+    }
+
+    /**
+     * Get config for restrictions depending on installed modules
+     *
+     * @return array Key - Module Id, Value - array of restrictions where key is type of possible restriction:
+     *         1. 'HumHubVersion'
+     *            'condition' - Sign to compare new installing HumHub version with allowed version for the modules, example: '>='
+     *            'version' - Allowed version, example: '1.10'
+     *            'message' - Error message in case of the restriction is applied
+     */
+    private function getModuleRestrictions(): array
+    {
+        return [
+            'enterprise' => [
+                'HumHubVersion' => [
+                    'condition' => '>=',
+                    'version' => '1.10',
+                    'message' => Yii::t('UpdaterModule.base', 'This HumHub version no longer supports the deprecated Enterprise Module. Please contact our support: {email}', [
+                        'email' => Link::to('hello@humhub.com', 'mailto:hello@humhub.com'),
+                    ])
+                ]
+            ]
+        ];
     }
 
     /**
@@ -140,6 +165,64 @@ class UpdatePackage
 
 
         return $notWritable;
+    }
+
+    /**
+     * Check if new HumHub version supports the current PHP version
+     *
+     * @return bool|string True - on supporting, String as new minimum supported PHP version
+     */
+    public function checkPhpVersion()
+    {
+        $newMinPhpVersion = $this->getNewConfigValue('minSupportedPhpVersion');
+        if ($newMinPhpVersion === null || version_compare(PHP_VERSION, $newMinPhpVersion, '>=')) {
+            return true;
+        }
+
+        return $newMinPhpVersion;
+    }
+
+    /**
+     * Check for restricted modules
+     *
+     * @return true|array TRUE - if no restriction, Array - error messages
+     */
+    public function checkRestrictedModules()
+    {
+        $errors = [];
+        foreach ($this->getModuleRestrictions() as $moduleId => $restrictions) {
+            if (!Yii::$app->getModule($moduleId)) {
+                continue;
+            }
+            foreach ($restrictions as $type => $restriction) {
+                switch ($type) {
+                    case 'HumHubVersion':
+                        if (version_compare($this->getNewConfigValue('version'), $restriction['version'], $restriction['condition'])) {
+                            $errors[] = $restriction['message'];
+                        }
+                        break;
+                }
+            }
+        }
+
+        return empty($errors) ? true : $errors;
+    }
+
+    private function getNewConfigValue($configVarName, $defaultValue = null)
+    {
+        $changedFiles = $this->getChangedFiles();
+
+        $configFileName = 'protected/humhub/config/common.php';
+        if (isset($changedFiles[$configFileName]['newFileMD5'])) {
+            $newConfigFilePath = $this->getNewFileDirectory() . DIRECTORY_SEPARATOR . $changedFiles[$configFileName]['newFileMD5'];
+            if (file_exists($newConfigFilePath)) {
+                if (preg_match('/[\'"]' . preg_quote($configVarName) . '[\'"]\s+=>\s+[\'"](.+?)[\'"]/', file_get_contents($newConfigFilePath), $match)) {
+                    return $match[1];
+                }
+            }
+        }
+
+        return $defaultValue;
     }
 
     public function install()