+++ messages/ja/libs_UpdatePackage.php
@@ -1,4 +1,4 @@
 <?php
 return array (
-  'Update download failed! (%error%)' => 'アップデートのダウンロードに失敗しました！ （％error％）',
+  'Update download failed! (%error%)' => 'アップデートのダウンロードに失敗しました！（%error%）',
 );